+++ source/Infrastructure/IncomingMessages/RequestAggregatedMeasureData/ValidateAggregatedMeasureDataRequestHandler.cs
@@ -1,0 +1,73 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Application.Configuration;
+using CimMessageAdapter.Messages;
+using CimMessageAdapter.Messages.RequestAggregatedMeasureData;
+using CimMessageAdapter.Response;
+using Domain.Actors;
+using Domain.ArchivedMessages;
+using Domain.Documents;
+using MediatR;
+using Receiver = CimMessageAdapter.Messages.RequestAggregatedMeasureData.RequestAggregatedMeasureDataReceiver;
+
+namespace Infrastructure.IncomingMessages.RequestAggregatedMeasureData;
+
+public class ValidateAggregatedMeasureDataRequestHandler
+    : IRequestHandler<ValidateAggregatedMeasureDataRequest, Result>
+{
+    private readonly RequestAggregatedMeasureDataReceiver _messageReceiver;
+    private readonly IArchivedMessageRepository _messageArchive;
+    private readonly ISystemDateTimeProvider _systemDateTimeProvider;
+
+    public ValidateAggregatedMeasureDataRequestHandler(
+        MessageParser messageParser,
+        Receiver messageReceiver,
+        ResponseFactory responseFactory,
+        ICorrelationContext correlationContext,
+        IArchivedMessageRepository messageArchive,
+        ISystemDateTimeProvider systemDateTimeProvider)
+    {
+        _messageReceiver = messageReceiver;
+        _messageArchive = messageArchive;
+        _systemDateTimeProvider = systemDateTimeProvider;
+    }
+
+    public async Task<Result> Handle(ValidateAggregatedMeasureDataRequest request, CancellationToken cancellationToken)
+    {
+        ArgumentNullException.ThrowIfNull(request);
+
+        var messageHeader = request.Request.IncomingMarketDocument?.Header;
+        ArgumentNullException.ThrowIfNull(messageHeader);
+
+        var timestamp = _systemDateTimeProvider.Now();
+
+        _messageArchive.Add(new ArchivedMessage(
+            messageHeader.MessageId,
+            DocumentType.RequestAggregatedMeasureData,
+            ActorNumber.Create(messageHeader.SenderId),
+            ActorNumber.Create(messageHeader.ReceiverId),
+            timestamp,
+            messageHeader.BusinessReason,
+            request.OriginalMessage));
+
+        var result = await _messageReceiver.ReceiveAsync(request.Request, cancellationToken)
+            .ConfigureAwait(false);
+
+        return result;
+    }
+}
