+++ source/Infrastructure/Transactions/AggregatedMeasureData/RequestedAggregatedMeasureDataConfiguration.cs
@@ -15,8 +15,10 @@
 using Application.IncomingMessages.RequestAggregatedMeasureData;
 using Application.Transactions.AggregatedMeasureData;
 using Application.Transactions.AggregatedMeasureData.Notifications;
+using CimMessageAdapter.Messages;
 using Domain.Transactions.AggregatedMeasureData;
 using Domain.Transactions.AggregatedMeasureData.Events;
+using Infrastructure.IncomingMessages.RequestAggregatedMeasureData;
 using Infrastructure.WholeSale;
 using MediatR;
 using Microsoft.Extensions.DependencyInjection;
@@ -29,6 +31,7 @@ internal static class RequestedAggregatedMeasureDataConfiguration
     {
         services.AddTransient<IRequestHandler<RequestAggregatedMeasureDataTransaction, Unit>, AggregatedMeasureDataRequestHandler>();
         services.AddTransient<IRequestHandler<NotifyWholesaleOfAggregatedMeasureDataRequest, Unit>, RequestAggregatedMeasuredDataFromWholesale>();
+        services.AddTransient<IRequestHandler<ValidateAggregatedMeasureDataRequest, Result>, ValidateAggregatedMeasureDataRequestHandler>();
         services.AddTransient<INotificationHandler<AggregatedMeasureProcessWasStarted>, NotifyWholesaleWhenAggregatedMeasureProcessWasStarted>();
         services.AddScoped<WholeSaleInbox>();
         services.AddScoped<IAggregatedMeasureDataProcessRepository, AggregatedMeasureDataProcessRepository>();
